+++ CHANGELOG.md
@@ -8,6 +8,7 @@ item (Added, Changed, Depreciated, Removed, Fixed, Security).
 
 - Updated: Refactor Horizontal Tabs block into a dynamic block. Add reordering functionality. 
 - Updated: Refactor Vertical Tabs block to dynamic block; allow sorting.
+- Updated: Admin menu order should now properly exclude predefined post types
 - Updated: Sticky Column block setting now respects masthead height.
 - Updated: Button block set to "Default" style now respects the "Width" setting.
 - Updated: Bump Block Editor Custom Alignments plugin to 1.1.3 to fix enqueuing warnings.